Refactor. Separated chart rendering components into different files
